refactor: interface 공용화 및 파일 위치 변경
