+++ src/components/common/CustomProForm/CustomProForm.style.tsx
@@ -1,0 +1,7 @@
+import { ProForm } from '@ant-design/pro-components';
+import styled from 'styled-components';
+
+// eslint-disable-next-line import/prefer-default-export
+export const ProFormWrap = styled(ProForm)`
+  flex-direction: column;
+  `;
